update close #110 1.10.22[2.7] 第-9.3周 小寒/大寒2-10

+++ Assets/Scripts/App/StartUp.cs
@@ -14,7 +14,7 @@ namespace App
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
         private static void DebugSetting()
         {
-            if (Debug.isDebugBuild && !Application.isEditor &&
+            if (false && Debug.isDebugBuild && !Application.isEditor &&
                 PlayerPrefs.GetString(Loading.VersionKey) != Application.version)
                 Directory.Delete(Application.persistentDataPath + "/com.unity.addressables", true);
